voss-server/voss/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
channels==4.0.0
daphne==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
colorama
groq
keyboard
numpy==1.26.4
# PyAudio
pydub
wave
openai
auraloss
speechbrain
sounddevice
torchvision==0.18.1
pysoundfile
gTTS
google-cloud-texttospeech
aiohttp